Wait for cluster client to be active
We had an issue (\#29) where the cluster client wasn't ready by the time
we got an observable reference.  This caused a null reference exception.

We fixed it by adding a Task.Delay before subscribing.  This commit
uses the built in lifecycle methods of orleans to expose a task which
lets a consumer wait until it is ready before using it.